Update to FastAPI 0.112 and Starlette 0.38
This commit will update/upgrade to
[FastAPI 0.112](https://fastapi.tiangolo.com/release-notes/)
and
[Starlette 0.38](https://www.starlette.io/release-notes/).

FastAPI 0.112 moves dependencies for the FastAPI CLI introduced in 0.111
to an optional extra, `fastapi[standard]`. Although the FastAPI release
notes consider this a breaking change, the FastAPI CLI was not used by
inboard, so FastAPI 0.112 will likely not be breaking for inboard users.

Starlette 0.38 makes various small changes, including removal of support
for an ASGI extension called
[path send](https://asgi.readthedocs.io/en/latest/extensions.html#path-send)
(aka "path-send" or `pathsend`) that was introduced in Starlette 0.36.
Support was removed because of issues with `BaseHTTPMiddleware`, though
note that `BaseHTTPMiddleware` may eventually be deprecated
([encode/starlette#2160](https://github.com/encode/starlette/discussions/2160),
[encode/starlette#2654](https://github.com/encode/starlette/discussions/2654)).
The Starlette release notes do not list this as a breaking change, but
it could be breaking for users who have started working with path send.