Simplify Gunicorn worker boot error test assertion
One of the Gunicorn worker tests verifies that Gunicorn shuts down
appropriately after a Uvicorn worker boot error. This test previously
had an assertion helper function, but the function was only used in
the worker boot test. The test also had some simple retry logic in which
an `AssertionError` would be caught and the assertion re-ran.

This commit will simplify the test by moving the assertion helper into
the test itself, and by using `Popen.wait()` instead of `time.sleep()`
for retries.