Fix 0.72.4 changelog generation
The changelog entry for inboard 0.72.4 revealed an issue with escaping
dunders in the changelog.

This repo has a GitHub Actions job to generate the changelog (b15efff).
The GitHub Actions changelog generation job includes a script to escape
"dunders" (double underscores, like `__init__`). Dunders make text bold
in Markdown when not escaped with backslashes or enclosed in `<code>`
elements with backticks. This was needed because there were some early
changelog entries that did not enclose dunders in backticks. `__init__`
would show up like "init" in bold because the Markdown parser saw it as
equivalent to `**init**`.

The changelog generation script has been correctly escaping some dunders
but is incorrectly escaping `AuthenticationMiddleware.__init__()`. This
commit will provide a short-term quick fix for the specific case from
0.72.4 (`AuthenticationMiddleware.__init__()`). A more robust fix should
be implemented for the longer term.