Fix changelog generator `$ESCAPE_DUNDERS` regex
This project includes a GitHub Actions job that generates a changelog
from Git tag messages and writes the output to Markdown documents
(b15efff). Some post-processing of these Git tag messages may be
necessary in order for them to render properly from the Markdown
documents. One reason for this post-processing is to handle "dunders"
(double underscores).

Dunder names like `__init__` have special meaning in Python.
https://docs.python.org/3/reference/datamodel.html#specialnames
https://docs.python.org/3/reference/lexical_analysis.html

Dunders also make text bold in Markdown unless escaped with backslashes
or enclosed in backticks. When the document is transformed to HTML, the
backticks are transformed to HTML code elements.

Some early changelog entries in this project did not enclose dunders in
backticks. When the text from these entries is formatted as Markdown and
rendered, text within dunders like `__init__` shows up as "init" in bold
because Markdown parsers read it as equivalent to `**init**`. To avoid
having these dunder names converted to bold text, the GitHub Actions job
runs a `sed` command to escape the dunders, like \_\_init\_\_.

The 0.72.4 changelog entry included mention of a Python dunder method,
`AuthenticationMiddleware.__init__()`. Although this text is enclosed in
backticks, the previous `sed` regex incorrectly escaped the underscores.
To mitigate this incorrect behavior, PR #121 (879782b) included a
short-term fix, simply un-escaping `AuthenticationMiddleware.__init__()`
after it had been incorrectly escaped by the old regex.

This commit will update the `$ESCAPE_DUNDERS` regex used in the `sed`
command so the `AuthenticationMiddleware.__init__()` text in the 0.72.4
changelog entry will not be escaped.