Fix changelog in separate GitHub Actions job step
This project includes a GitHub Actions job that generates a changelog
from Git tag messages and writes the output to Markdown documents
(b15efff). Some post-processing of these Git tag messages may be
necessary in order for them to render properly from the Markdown
documents. This post-processing is performed with `sed` commands. The
`sed` commands are in the same step as the changelog generation
commands, making the step difficult to read and making step failures
more difficult to debug.

This commit will move the `sed` commands to a separate GitHub Actions
job step, and will add comments to the new step explaining why the step
is needed.