Update to FastAPI 0.116 and Starlette 0.47
This commit will update/upgrade to
[FastAPI 0.116](https://fastapi.tiangolo.com/release-notes/) and
[Starlette 0.47](https://www.starlette.io/release-notes/).
This is a minor release to align with FastAPI and Starlette versioning.

FastAPI 0.116 introduces support for deploying to FastAPI Cloud with
the `fastapi deploy` CLI command.

Starlette 0.47 reintroduces support for an ASGI extension called
[path send](https://asgi.readthedocs.io/en/latest/extensions.html#path-send)
(aka "path-send" or `pathsend`). It was introduced in Starlette 0.36,
then removed in 0.38 because of issues with `BaseHTTPMiddleware`. Now
they're giving it another try.