Remove deprecated `show_error_codes` mypy setting
Mypy now shows error codes by default. It's unclear when this changed.
It doesn't seem to be mentioned in the changelog, and both the original
`show_error_codes` setting and the deprecation seem to be undocumented.

This commit will remove `show_error_codes`/`--show-error-codes`.

The inverse `hide_error_codes`/`--hide-error-codes` is still supported.

https://github.com/python/mypy/blob/master/CHANGELOG.md
https://github.com/python/mypy/pull/13542
https://github.com/python/mypy/issues/17083
https://mypy.readthedocs.io/en/stable/error_codes.html

https://github.com/SchemaStore/schemastore/pull/3422
https://github.com/SchemaStore/schemastore/issues/3467
https://github.com/SchemaStore/schemastore/pull/3472
https://www.schemastore.org/partial-mypy.json