Use `$HATCH_ENV_TYPE_VIRTUAL_PATH` consistently
The Dockerfile includes `HATCH_ENV_TYPE_VIRTUAL_PATH=.venv` so that the
virtual environment is installed to a predictable path, but the rest of
the repo used `path = ".venv"` in `pyproject.toml` to set the path.

This commit will move Hatch virtualenv paths from `pyproject.toml` to
the environment variable `$HATCH_ENV_TYPE_VIRTUAL_PATH`. This allows
users to customize the path to their liking, and also allows the GitHub
Actions workflow jobs to use `$HATCH_ENV_TYPE_VIRTUAL_PATH` to check the
virtual environment path, instead of having to hard-code `".venv"`.

https://hatch.pypa.io/latest/plugins/environment/virtual/