Update to FastAPI 0.121.3 and Starlette 0.50
This commit will update/upgrade to
[FastAPI 0.121.3](https://fastapi.tiangolo.com/release-notes/) and
[Starlette 0.50](https://starlette.dev/release-notes/).

FastAPI 0.121.3 adds support for Starlette 0.50 and provides some other
small changes and internal refactors.

Starlette 0.50 drops Python 3.9 support. This is a breaking change for
Python 3.9 users. Note that, while Starlette doesn't support Python 3.9,
[FastAPI still does](https://github.com/fastapi/fastapi/blob/4f3ff79/pyproject.toml#L37-L38).
This doesn't make much sense because FastAPI depends on Starlette for
much of its functionality. FastAPI probably shouldn't be using Starlette
with unsupported versions of Python.

There are some issues with Starlette's type annotations (as usual), so
some of the BasedPyright ignores will be updated and the baseline file
will be updated (`basedpyright --writebaseline`).