Update to FastAPI 0.122
This commit will update/upgrade to
[FastAPI 0.122](https://fastapi.tiangolo.com/release-notes/).

FastAPI 0.122 updates the behavior of `fastapi.security` classes. These
classes will now return `401 Unauthorized` instead of `403 Forbidden`.
See the
[HTTP Basic auth docs](https://fastapi.tiangolo.com/advanced/security/http-basic-auth/)
for general info, and the related
[how-to](https://fastapi.tiangolo.com/how-to/authentication-error-status-code/)
if custom status codes are needed for authorization errors.

Note that this may be a breaking change for some projects if application
logic or tests were hard-coded to expect a `401` status code. inboard
anticipated this situation years ago, so the tests of HTTP Basic auth
were written to allow either a `401` or `403` for these errors as seen
[here](https://github.com/br3ndonland/inboard/blob/b429f5c/tests/app/test_main.py#L162).